Move some beans to web configuration where they belong
The beans in question are:

- message sources
- character encoding filter

Also remove auto-wiring of `OsiamHome`, because it's not needed anymore.
