+++ src/main/java/org/osiam/Osiam.java
@@ -6,23 +6,18 @@ import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
 import org.flywaydb.core.Flyway;
 import org.flywaydb.core.api.MigrationVersion;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.context.web.SpringBootServletInitializer;
-import org.springframework.context.MessageSource;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.EnableAspectJAutoProxy;
 import org.springframework.context.annotation.Primary;
-import org.springframework.context.support.ReloadableResourceBundleMessageSource;
 import org.springframework.security.authentication.encoding.ShaPasswordEncoder;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
-import org.springframework.web.filter.CharacterEncodingFilter;
 
-import javax.servlet.Filter;
 import javax.sql.DataSource;
 import java.util.Map;
 
@@ -53,9 +48,6 @@ public class Osiam extends SpringBootServletInitializer {
     @Value("${org.osiam.db.vendor}")
     private String databaseVendor;
 
-    @Autowired
-    private OsiamHome osiamHome;
-
     public static void main(String[] args) {
         SpringApplication application = new SpringApplication(Osiam.class);
         application.setDefaultProperties(DEFAULT_PROPERTIES);
@@ -66,14 +58,6 @@ public class Osiam extends SpringBootServletInitializer {
     protected SpringApplicationBuilder configure(SpringApplicationBuilder applicationBuilder) {
         applicationBuilder.application().setDefaultProperties(DEFAULT_PROPERTIES);
         return applicationBuilder.sources(Osiam.class);
-    }
-
-    @Bean
-    public Filter characterEncodingFilter() {
-        CharacterEncodingFilter characterEncodingFilter = new CharacterEncodingFilter();
-        characterEncodingFilter.setEncoding("UTF-8");
-        characterEncodingFilter.setForceEncoding(true);
-        return characterEncodingFilter;
     }
 
     @Primary
@@ -103,14 +87,5 @@ public class Osiam extends SpringBootServletInitializer {
         ShaPasswordEncoder passwordEncoder = new ShaPasswordEncoder(512);
         passwordEncoder.setIterations(1000);
         return passwordEncoder;
-    }
-
-    @Bean
-    public MessageSource messageSource() {
-        ReloadableResourceBundleMessageSource messageSource = new ReloadableResourceBundleMessageSource();
-        messageSource.setBasename(osiamHome.getI18nDirectory() + "/login");
-        messageSource.setDefaultEncoding("utf-8");
-        messageSource.setCacheSeconds(-1);
-        return messageSource;
     }
 }
